+++ actions/src/sema4ai/actions/__init__.py
@@ -42,7 +42,7 @@ from ._response import ActionError, Response
 from ._secret import OAuth2Secret, Secret
 from ._table import Row, RowValue, Table
 
-__version__ = "1.3.2"
+__version__ = "1.3.3"
 version_info = [int(x) for x in __version__.split(".")]